Fix setuptools issues
syntax error

syntax error

Fix a setup issue

Fix a setup issue

Fix a setup issue

Fix a setup issue

Fix a setup issue

Fix a setup issue
